+++ Stitch/Graph/StitchAI/GraphPrompting/Model/RequestTypes/AICodeGenRequest.swift
@@ -7,129 +7,6 @@
 
 import SwiftUI
 
-protocol StitchAIFunctionRequestable: StitchAIRequestable where
-Self.InitialDecodedResult == [OpenAIToolCallResponse], Self.InitialDecodedResult == Self.FinalDecodedResult, Self.Body: StitchAIRequestableFunctionBody { }
-
-extension StitchAIFunctionRequestable {
-    var functionName: String { self.body.functionName }
-}
-
-extension StitchAIFunctionRequestable {
-    func decodeMessage<ResultType>(from message: OpenAIMessage,
-                                   document: StitchDocumentViewModel,
-                                   aiManager: StitchAIManager,
-                                   resultType: ResultType.Type) throws -> ResultType where Self.InitialDecodedResult == [OpenAIToolCallResponse], Self.InitialDecodedResult == Self.FinalDecodedResult, ResultType: Decodable, Self.Body: StitchAIRequestableFunctionBody {
-        let toolsResponse = try Self.parseOpenAIResponse(message: message)
-        
-        guard let tool = toolsResponse.first?.function,
-              tool.name == self.body.functionName,
-              let swiftUISourceCodeData = tool.arguments.data(using: .utf8) else {
-            throw StitchAIManagerError.functionDecodingFailed
-        }
-        
-        let decodedResult = try JSONDecoder().decode(
-            resultType.self,
-            from: swiftUISourceCodeData
-        )
-        
-        return decodedResult
-    }
-}
-
-// TODO: move
-// TODO: actually move
-struct AICodeEditRequest: StitchAIFunctionRequestable {
-    let id: UUID
-    let userPrompt: String             // User's input prompt
-    let config: OpenAIRequestConfig // Request configuration settings
-    let body: AICodeEditBody_V0.AICodeEditRequestBody
-    static let willStream: Bool = false
-    
-    init(id: UUID,
-         prompt: String,
-         toolMessages: [OpenAIMessage],
-         systemPrompt: String,
-         config: OpenAIRequestConfig = .default) throws {
-        
-        // The id of the user's inference call; does not change across retries etc.
-        self.id = id
-        
-        self.userPrompt = prompt
-        self.config = config
-        
-        // Construct http payload
-        self.body = try AICodeEditBody_V0.AICodeEditRequestBody(
-            userPrompt: prompt,
-            toolMessages: toolMessages,
-            systemPrompt: systemPrompt)
-    }
-    
-    @MainActor
-    func willRequest(document: StitchDocumentViewModel,
-                     canShareData: Bool,
-                     requestTask: Self.RequestTask) {
-        // Nothing to do
-    }
-    
-    static func validateResponse(decodedResult: [OpenAIToolCallResponse]) throws -> [OpenAIToolCallResponse] {
-        decodedResult
-    }
-    
-    @MainActor
-    func onSuccessfulDecodingChunk(result: [OpenAIToolCallResponse],
-                                   currentAttempt: Int) {
-        fatalErrorIfDebug()
-    }
-    
-    static func buildResponse(from streamingChunks: [[OpenAIToolCallResponse]]) throws -> [OpenAIToolCallResponse] {
-        // Unsupported
-        fatalError()
-    }
-}
-
-protocol StitchAIGraphBuilderRequestable: StitchAIFunctionRequestable {
-    // TODO: Group the entire code gen/edit requests under one handler
-    
-    static var type: StitchAIRequestBuilder_V0.StitchAIRequestType { get }
-        
-    func createCode(document: StitchDocumentViewModel,
-                    aiManager: StitchAIManager,
-                    systemPrompt: String) async throws -> (String, OpenAIMessage)
-}
-
-extension StitchAIGraphBuilderRequestable {
-    @MainActor
-    static func systemPrompt(graph: GraphState) throws -> String {
-        try StitchAIManager
-            .stitchAIGraphBuilderSystem(graph: graph,
-                                        requestType: Self.type)
-    }
-}
-
-extension StitchAIGraphBuilderRequestable {
-    @MainActor
-    func willRequest(document: StitchDocumentViewModel,
-                     canShareData: Bool,
-                     requestTask: Self.RequestTask) {
-        // Nothing to do
-    }
-    
-    static func validateResponse(decodedResult: [OpenAIToolCallResponse]) throws -> [OpenAIToolCallResponse] {
-        decodedResult
-    }
-    
-    @MainActor
-    func onSuccessfulDecodingChunk(result: [OpenAIToolCallResponse],
-                                   currentAttempt: Int) {
-        fatalErrorIfDebug()
-    }
-    
-    static func buildResponse(from streamingChunks: [[OpenAIToolCallResponse]]) throws -> [OpenAIToolCallResponse] {
-        // Unsupported
-        fatalError()
-    }
-}
-
 struct AICodeGenFromGraphRequest: StitchAIGraphBuilderRequestable {
     static let type = StitchAIRequestBuilder_V0.StitchAIRequestType.userPrompt
     
@@ -194,11 +71,6 @@ struct AICodeGenFromGraphRequest: StitchAIGraphBuilderRequestable {
         return (swiftUIEditedCode, msgFromEditCodeRequest)
     }
 }
-
-
-
-// TODO: MOVE
-
 
 struct AICodeGenFromImageRequest: StitchAIGraphBuilderRequestable {
     static let type = StitchAIRequestBuilder_V0.StitchAIRequestType.imagePrompt
